rename telemetry to analytics
