+++ docs/learn/tests/helpers.bash
@@ -10,9 +10,9 @@ common_setup() {
     #   otherwise infinite recursion when DAGGER_BINARY is not set.
     export DAGGER="${DAGGER_BINARY:-$(bash -c 'command -v dagger')}"
 
-    # Disable telemetry
-    DAGGER_TELEMETRY_DISABLE="1"
-    export DAGGER_TELEMETRY_DISABLE
+    # Disable analytics
+    DAGGER_DISABLE_ANALYTICS="1"
+    export DAGGER_DISABLE_ANALYTICS
 
     # Set the project to the universe directory (so tests can run from anywhere)
     UNIVERSE="$( cd "$( dirname "$BATS_TEST_FILENAME" )" >/dev/null 2>&1 && pwd )"
